Rename to user-interface-text.ts
