+++ ui/dashboards/src/constants/index.ts
@@ -11,4 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './copy';
+export * from './user-interface-text';
