+++ ui/dashboards/src/constants/user-interface-text.ts
